Move MarkdownText component
